pyproject.toml
[project]
name = "octopi-demo-mcp"
version = "0.1.1"
description = "Octopi Demo MCP for scanning vulnerabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0"
]
